+++ 01-MCAL/02-GPIO_DRIVER/GPIO.h
@@ -120,7 +120,7 @@ extern ERROR_STATUS GPIO_writePin(GPIO_t * peri, u32 value);
   Output: ERROR_STATUS
 
  */
-extern ERROR_STATUS GPIO_directWritePin(void * port ,u32 pin, u32 value);
+extern ERROR_STATUS GPIO_directWritePin(void * port ,u32 pin, u8 value);
 
 /* 
   Description: This function shall read value of a pin 
@@ -149,6 +149,6 @@ extern ERROR_STATUS GPIO_readPin(GPIO_t * peri, u32 *value);
   Output: ERROR_STATUS
 
  */
-extern ERROR_STATUS GPIO_directReadPin(void * port ,u32 pin, u32 * value);
+extern ERROR_STATUS GPIO_directReadPin(void * port ,u32 pin, u8 * value);
 
 #endif
